requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.27
asyncmy==0.2.9
cryptography==42.0.5  # Bu satırı ekleyin veya güncelleyin
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.1
playwright==1.42.0